+++ Source By Mr Blue/src/nro/models/mob_bigboss/Hirudegarn.java
@@ -191,7 +191,7 @@ public class Hirudegarn extends BigBoss {
                             msg.writer().writeByte(1);
                             int dame = player.injured(null, this.point.getDameAttack(), false, true);
                             msg.writer().writeInt((int) player.id);
-                            msg.writer().writeInt(dame);
+                            msg.writer().writeLong(dame);
                             break;
                         case 3:
                             this.location.x = (short) player.location.x;
@@ -204,7 +204,7 @@ public class Hirudegarn extends BigBoss {
                                 Player pl = this.zone.getPlayers().get(i);
                                 dame = pl.injured(null, this.point.getDameAttack(), false, true);
                                 msg.writer().writeInt((int) pl.id);
-                                msg.writer().writeInt(dame);
+                                msg.writer().writeLong(dame);
                             }
                             break;
                     }
